# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Host Admin API (Stub)
  version: 0.1.5
servers:
  - url: http://localhost:8080
paths:
  /runs:
    post:
      summary: Startet einen neuen Agenten-Run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/run.request.schema.json'
      responses:
        '200':
          description: Run angenommen
        '400':
          description: Ungültige Eingabe (invalid_argument, invalid_chain_spec, invalid_chain_version, invalid_chain)
...
  /runs/{id}:
    get:
      summary: Liefert Run-Trace inkl. Events und Chain-Kontext
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run-Details
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  agent_id:
                    type: string
                    nullable: true
                  task_id:
                    type: string
                    nullable: true
                  chain_id:
                    type: string
                    nullable: true
                  chain_version_id:
                    type: string
                    nullable: true
                  events:
                    type: array
                    items:
                      type: object
                  output:
                    type: string
                    nullable: true
        '400':
          description: Ungültige Eingabe (invalid_run_id)
        '404':
          description: Run nicht gefunden
...